fix(slider): prevent circular dependecies
Closes #5075
