+++ projects/igniteui-angular/src/lib/slider/index.ts
@@ -1,0 +1,2 @@
+export * from './slider.component';
+export * from './slider.common';
