+++ projects/igniteui-angular/src/lib/slider/slider.component.ts
@@ -3,7 +3,6 @@ import {
     AfterViewInit, Component, ElementRef, EventEmitter,
     HostBinding, Input, NgModule, OnInit, Output, Renderer2,
     ViewChild,
-    Directive,
     TemplateRef,
     ContentChild,
     AfterContentInit,
@@ -15,67 +14,17 @@ import {
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { EditorProvider } from '../core/edit-provider';
 import { DeprecateProperty } from '../core/deprecateDecorators';
-import { IgxSliderThumbModule, IgxSliderThumbComponent } from './thumb/thumb-slider.component';
-import { Subject, merge, concat } from 'rxjs';
+import { IgxSliderThumbComponent } from './thumb/thumb-slider.component';
+import { Subject, merge } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
+import { SliderHandle,
+    IgxThumbFromTemplateDirective,
+    IgxThumbToTemplateDirective,
+    IRangeSliderValue,
+    SliderType,
+    ISliderValueChangeEventArgs
+} from './slider.common';
 
-/**
- * Template directive that allows you to set a custom template representing the lower label value of the {@link IgxSliderComponent}
- *
- *```html
- * <igx-slider>
- *  <ng-template igxSliderThumbFrom let-value let-labels>{{value}}</ng-template>
- * </igx-slider>
- * ```
- *
- * @context {@link IgxSliderComponent.context}
- */
-@Directive({
-    selector: '[igxSliderThumbFrom]'
-})
-export class IgxThumbFromTemplateDirective {}
-
-/**
- * Template directive that allows you to set a custom template representing the upper label value of the {@link IgxSliderComponent}
- *
- * ```html
- * <igx-slider>
- *  <ng-template igxSliderThumbTo let-value let-labels>{{value}}</ng-template>
- * </igx-slider>
- * ```
- *
- * @context {@link IgxSliderComponent.context}
- */
-@Directive({
-    selector: '[igxSliderThumbTo]'
-})
-export class IgxThumbToTemplateDirective {}
-
-export enum SliderType {
-    /**
-     * Slider with single thumb.
-     */
-    SLIDER,
-    /**
-     *  Range slider with multiple thumbs, that can mark the range.
-     */
-    RANGE
-}
-
-export enum SliderHandle {
-    FROM,
-    TO
-}
-
-export interface IRangeSliderValue {
-    lower: number;
-    upper: number;
-}
-
-export interface ISliderValueChangeEventArgs {
-    oldValue: number | IRangeSliderValue;
-    value: number | IRangeSliderValue;
-}
 
 const noop = () => {
 };
@@ -1240,9 +1189,9 @@ export class IgxSliderComponent implements
  * @hidden
  */
 @NgModule({
-    declarations: [IgxSliderComponent, IgxThumbFromTemplateDirective, IgxThumbToTemplateDirective],
-    exports: [IgxSliderComponent, IgxThumbFromTemplateDirective, IgxThumbToTemplateDirective],
-    imports: [CommonModule, IgxSliderThumbModule]
+    declarations: [IgxSliderComponent, IgxThumbFromTemplateDirective, IgxThumbToTemplateDirective, IgxSliderThumbComponent],
+    exports: [IgxSliderComponent, IgxThumbFromTemplateDirective, IgxThumbToTemplateDirective, IgxSliderThumbComponent],
+    imports: [CommonModule]
 })
 export class IgxSliderModule {
 }
