+++ src/app/slider/slider.sample.html
@@ -13,7 +13,7 @@
 
             <article class="sample-column">
                 <h4 class="sample-title">Range slider</h4>
-                <igx-slider #slider2 [type]="sliderType" [disabled]="true" [(ngModel)]="rangeValue" [minValue]="0" [maxValue]="100">
+                <igx-slider #slider2 [type]="sliderType" [disabled]="false" [(ngModel)]="rangeValue" [minValue]="0" [maxValue]="100">
                 </igx-slider>
             </article>
